Bullet inlining
