+++ main/src/unity/content/blocks/KoruhBlocks.java
@@ -9,6 +9,7 @@ import mindustry.*;
 import mindustry.content.*;
 import mindustry.entities.*;
 import mindustry.entities.bullet.*;
+import mindustry.entities.effect.*;
 import mindustry.entities.pattern.*;
 import mindustry.gen.*;
 import mindustry.graphics.*;
@@ -37,6 +38,7 @@ import static arc.Core.bundle;
 import static mindustry.Vars.tilesize;
 import static mindustry.type.ItemStack.empty;
 import static mindustry.type.ItemStack.with;
+import static unity.content.UnityStatusEffects.*;
 
 public class KoruhBlocks {
     public static final Category expCategory = Category.distribution;
@@ -561,7 +563,14 @@ public class KoruhBlocks {
             shootSound = Sounds.laser;
 
             powerUse = 7f;
-            shootType = UnityBullets.laser;
+            shootType = new ExpLaserBulletType(150f, 30f){{
+                damageInc = 7f;
+                status = StatusEffects.shocked;
+                statusDuration = 3 * 60f;
+                expGain = buildingExpGain = 2;
+                fromColor = Pal.accent;
+                toColor = Pal.lancerLaser;
+            }};
 
             maxLevel = 10;
             expFields = new EField[]{
@@ -593,7 +602,48 @@ public class KoruhBlocks {
             heatColor = Color.red;
             shootSound = Sounds.laser;
 
-            shootType = UnityBullets.shardLaser;
+            shootType = new ExpLaserBulletType(150f, 30f){{
+                status = StatusEffects.shocked;
+                statusDuration = 3 * 60f;
+                fragBullet = new ExpBasicBulletType(2f, 10f){
+                    {
+                        lifetime = 20f;
+                        pierceCap = 10;
+                        pierceBuilding = true;
+                        backColor = Color.white.cpy().lerp(Pal.lancerLaser, 0.1f);
+                        frontColor = Color.white;
+                        hitEffect = Fx.none;
+                        despawnEffect = Fx.none;
+                        smokeEffect = Fx.hitLaser;
+                        hittable = false;
+                        reflectable = false;
+                        lightColor = Color.white;
+                        lightOpacity = 0.6f;
+
+                        expChance = 0.15f;
+                        fromColor = Pal.lancerLaser;
+                        toColor = Pal.sapBullet;
+                    }
+
+                    @Override
+                    public void draw(Bullet b){
+                        Draw.color(getColor(b));
+                        Lines.stroke(2f * b.fout(0.7f) + 0.01f);
+                        Lines.lineAngleCenter(b.x, b.y, b.rotation(), 8f);
+                        Lines.stroke(1.3f * b.fout(0.7f) + 0.01f);
+                        Draw.color(frontColor);
+                        Lines.lineAngleCenter(b.x, b.y, b.rotation(), 5f);
+                        Draw.reset();
+                    }
+                };
+
+                expGain = buildingExpGain = 2;
+                damageInc = 5f;
+                fromColor = Pal.lancerLaser;
+                toColor = Pal.sapBullet;
+
+                chargeEffect = new MultiEffect(UnityFx.laserCharge, UnityFx.laserChargeBegin);
+            }};
 
             maxLevel = 30;
             expFields = new EField[]{
@@ -605,7 +655,36 @@ public class KoruhBlocks {
         }};
 
         laserFrost = new ExpLiquidTurret("frost-laser-turret"){{
-            ammo(Liquids.cryofluid, UnityBullets.frostLaser);
+            ammo(Liquids.cryofluid, new ExpLaserBulletType(170f, 130f){
+                {
+                    status = StatusEffects.freezing;
+                    statusDuration = 3 * 60f;
+                    shootEffect = UnityFx.shootFlake;
+
+                    expGain = 2;
+                    buildingExpGain = 3;
+                    damageInc = 2.5f;
+                    fromColor = Liquids.cryofluid.color;
+                    toColor = Color.cyan;
+                    blip = true;
+                }
+
+                @Override
+                public void handleExp(Bullet b, float x, float y, int amount){
+                    super.handleExp(b, x, y, amount);
+                    freezePos(b, x, y);
+                }
+
+                public void freezePos(Bullet b, float x, float y){
+                    int lvl = getLevel(b);
+                    float rad = 3.5f;
+                    UnityFx.freezeEffect.at(x, y, lvl / rad + 10f, getColor(b));
+                    UnitySounds.laserFreeze.at(x, y);
+
+                    Damage.status(b.team, x, y, 10f + lvl / rad, status, 60f + lvl * 6f, true, true);
+                    Damage.status(b.team, x, y, 10f + lvl / rad, UnityStatusEffects.disabled, 2f * lvl, true, true);
+                }
+            });
             requirements(Category.turret, with(UnityItems.denseAlloy, 20, Items.metaglass, 15));
             size = 2;
             health = 1000;
@@ -628,7 +707,9 @@ public class KoruhBlocks {
             size = 3;
             health = 2000;
 
-            reload = UnityBullets.distField.lifetime / 3f;
+            float distFieldLife = 6 * 60;
+
+            reload = distFieldLife / 3f;
             coolantMultiplier = 2f;
             range = 140f;
 
@@ -648,11 +729,70 @@ public class KoruhBlocks {
             fromColor = UnityPal.lancerSap3;
             toColor = Pal.place;
 
-            shootType = UnityBullets.fractalLaser;
+            shootType = new ExpLaserFieldBulletType(170f, 130f){{
+                damageInc = 6f;
+                lengthInc = 2f;
+                fields = 2;
+                fieldInc = 0.15f;
+                width = 2;
+                expGain = buildingExpGain = 1;
+                fromColor = Pal.lancerLaser.cpy().lerp(Pal.place, 0.5f);
+                toColor = Pal.place;
+                maxRange = 150f + 2f * 30f; //Account for range increase
+
+                distField = new DistFieldBulletType(0, -1){{
+                    centerColor = Pal.lancerLaser.cpy().a(0);
+                    edgeColor = Pal.place;
+                    distSplashFx = UnityFx.distSplashFx;
+                    distStart = UnityFx.distStart;
+                    distStatus = distort;
+
+                    collidesTiles = false;
+                    collides = false;
+                    collidesAir = false;
+                    keepVelocity = false;
+
+                    lifetime = distFieldLife;
+                    radius = 3f*8;
+                    radiusInc = 0.1f*8;
+                    bulletSlow = 0.1f;
+                    bulletSlowInc = 0.025f;
+                    damageLimit = 100f;
+                    distDamage = 0.1f;
+                    expChance = 0.5f/60;
+                    expGain = 1;
+                }};
+
+
+                smallDistField = new DistFieldBulletType(0, -1){{
+                    centerColor = Pal.lancerLaser.cpy().a(0);
+                    edgeColor = Pal.place;
+                    distSplashFx = UnityFx.distSplashFx;
+                    distStart = UnityFx.distStart;
+                    distStatus = distort;
+
+                    collidesTiles = false;
+                    collides = false;
+                    collidesAir = false;
+                    keepVelocity = false;
+
+                    lifetime = 2.5f * 60;
+                    radius = 1.5f*8;
+                    radiusInc = 0;
+                    bulletSlow = 0.05f;
+                    bulletSlowInc = 0;
+                    damageLimit = 50f;
+                    distDamage = 0.05f;
+                    expChance = 0.1f/60;
+                    expGain = 1;
+                }};
+
+                chargeEffect = new MultiEffect(UnityFx.laserFractalCharge, UnityFx.laserFractalChargeBegin);
+            }};
 
             maxLevel = 30;
             expFields = new EField[]{
-                    new LinearReloadTime(v -> reload = v, UnityBullets.distField.lifetime / 3f, -2f),
+                    new LinearReloadTime(v -> reload = v, distFieldLife / 3f, -2f),
                     new EField.ELinear(v -> range = v, 140f, 0.25f * tilesize, Stat.shootRange, v -> Strings.autoFixed(v / tilesize, 2) + " blocks")
             };
 
@@ -682,7 +822,60 @@ public class KoruhBlocks {
             heatColor = Color.red;
             fromColor = UnityPal.lancerSap3;
             shootSound = Sounds.plasmaboom;
-            shootType = UnityBullets.branchLaser;
+            shootType = new ExpLaserBulletType(140f, 20f){{
+                status = StatusEffects.shocked;
+                statusDuration = 3 * 60f;
+                fragBullets = 3;
+                fragBullet = new ExpBulletType(3.5f, 15f){
+                    {
+                        trailWidth = 2f;
+                        weaveScale = 0.6f;
+                        weaveMag = 0.5f;
+                        homingPower = 0.4f;
+                        lifetime = 30f;
+                        shootEffect = Fx.hitLancer;
+                        hitEffect = despawnEffect = HitFx.branchFragHit;
+                        pierceCap = 10;
+                        pierceBuilding = true;
+                        splashDamageRadius = 4f;
+                        splashDamage = 4f;
+                        status = UnityStatusEffects.plasmaed;
+                        statusDuration = 180f;
+                        trailLength = 6;
+                        trailColor = Color.white;
+
+                        fromColor = Pal.lancerLaser.cpy().lerp(Pal.sapBullet, 0.5f);
+                        toColor = Pal.sapBullet;
+                        expGain = 1;
+                        expOnHit = true;
+                    }
+
+                    @Override
+                    public void init(){
+                        super.init();
+                        despawnHit = false;
+                    }
+
+                    @Override
+                    public void draw(Bullet b){
+                        drawTrail(b);
+
+                        Draw.color(getColor(b));
+                        Fill.square(b.x, b.y, trailWidth, b.rotation() + 45);
+                        Draw.color();
+                    }
+                };
+                maxRange = 150f + 2f * 30f; //Account for range increase
+
+                expGain = buildingExpGain = 1;
+                damageInc = 6f;
+                lengthInc = 2f;
+                fromColor = Pal.lancerLaser.cpy().lerp(Pal.sapBullet, 0.5f);
+                toColor = Pal.sapBullet;
+                hitMissed = true;
+
+                chargeEffect = new MultiEffect(UnityFx.laserChargeShort, UnityFx.laserChargeBegin);
+            }};
 
             shootY = size * tilesize / 2.7f;
             shoot.shots = 4;
@@ -715,7 +908,9 @@ public class KoruhBlocks {
             shootSound = Sounds.laser;
 
             shootType = new GeyserLaserBulletType(185f, 30f){{
-                geyser = UnityBullets.laserGeyser;
+                geyser = new GeyserBulletType(){{
+                    damageInc = 2f;
+                }};
                 damageInc = 5f;
                 maxRange = 185f;
             }};
@@ -749,7 +944,21 @@ public class KoruhBlocks {
             toColor = UnityPal.exp;
             shootSound = Sounds.laserblast;
             chargeSound = Sounds.lasercharge;
-            shootType = UnityBullets.breakthroughLaser;
+            shootType = new ExpLaserBlastBulletType(500f, 1200f){{
+                damageInc = 1000f;
+                lengthInc = 150f;
+                largeHit = true;
+                width = 80f;
+                widthInc = 10f;
+                lifetime = 65f;
+                lightningSpacingInc = -5f;
+                lightningDamageInc = 30f;
+                hitUnitExpGain = 1;
+                hitBuildingExpGain = 1;
+                sideLength = 0f;
+                sideWidth = 0f;
+                chargeEffect = UnityFx.laserBreakthroughChargeBegin;
+            }};
 
             maxLevel = 1;
             expScale = 30;
@@ -786,8 +995,38 @@ public class KoruhBlocks {
                         damage = 4;
                         drag = 0.01f;
                     }},
-                    Items.coal, UnityBullets.coalBlaze,
-                    Items.pyratite, UnityBullets.pyraBlaze
+                    Items.coal, new ExpBulletType(3.35f, 32f){{
+                        ammoMultiplier = 3;
+                        hitSize = 7f;
+                        lifetime = 24f;
+                        pierce = true;
+                        statusDuration = 60 * 4f;
+                        shootEffect = ShootFx.shootSmallBlaze;
+                        hitEffect = Fx.hitFlameSmall;
+                        despawnEffect = Fx.none;
+                        status = StatusEffects.burning;
+                        keepVelocity = true;
+                        hittable = false;
+
+                        expOnHit = true;
+                        expChance = 0.5f;
+                    }},
+                    Items.pyratite, new ExpBulletType(3.35f, 46f){{
+                        ammoMultiplier = 3;
+                        hitSize = 7f;
+                        lifetime = 24f;
+                        pierce = true;
+                        statusDuration = 60 * 4f;
+                        shootEffect = ShootFx.shootPyraBlaze;
+                        hitEffect = Fx.hitFlameSmall;
+                        despawnEffect = Fx.none;
+                        status = StatusEffects.burning;
+                        keepVelocity = false;
+                        hittable = false;
+
+                        expOnHit = true;
+                        expChance = 0.6f;
+                    }}
             );
 
             size = 3;
